requirements.txt
argparse
cache_to_disk
msgspec
numpy
pandas
requests
platformdirs
pyqtgraph
PySide6
pulp
gekko
scipy
pillow